refactor(tui): split help overlay content out of ui::overlay
ui::overlay crossed the 1000-line file-size warn threshold once the
update-available popup landed. Splits the larger, independent `?` help
overlay (keymap/markers/glossary content and its own draw function)
into a new ui::help_overlay module, leaving ui::overlay with just the
small popups (jump-target, update-available) and their shared
centered_rect layout primitive. Pure move — no behavior change, same
886 tests passing before and after.
