+++ rinkaku-tui/src/ui/mod.rs
@@ -13,6 +13,7 @@ mod blast_radius;
 mod detail_pane;
 mod diff_pane;
 mod entry;
+mod help_overlay;
 mod overlay;
 mod review_overlay;
 mod scroll;
@@ -25,7 +26,8 @@ use crate::diff_view::FileHunks;
 use crate::highlight::HighlightedFile;
 use crate::source::HighlightedSourceView;
 use entry::draw_entry_screen;
-use overlay::{draw_help_overlay, draw_jump_popup, draw_update_prompt};
+use help_overlay::draw_help_overlay;
+use overlay::{draw_jump_popup, draw_update_prompt};
 use ratatui::Frame;
 use ratatui::layout::{Constraint, Layout, Rect};
 use rinkaku_core::render::Report;
